+++ lib/mongoid/userstamp/user.rb
@@ -1,41 +1,39 @@
 # -*- encoding : utf-8 -*-
 module Mongoid
-  module Userstamp
-    module User
-      extend ActiveSupport::Concern
+module Userstamp
 
-      included do
-        cattr_accessor :user_key do
-          "mongoid_userstamp/#{self.to_s.underscore}".to_sym
-        end
+  module User
 
-        def current?
-          !Thread.current[user_key].nil? && self._id == Thread.current[user_key]._id
-        end
+    extend ActiveSupport::Concern
+
+    included do
+
+      def current?
+        self._id == Mongoid::Userstamp.current_user(self.class).try(:_id)
+      end
+    end
+
+    module ClassMethods
+
+      def current
+        Mongoid::Userstamp.current_user(self)
       end
 
-      module ClassMethods
-        def current
-          Thread.current[user_key]
+      def current=(value)
+        Mongoid::Userstamp.current_user = value
+      end
+
+      def do_as(user, &block)
+        old = self.current
+        begin
+          self.current = user
+          response = block.call unless block.nil?
+        ensure
+          self.current = old
         end
-
-        def current=(value)
-          Thread.current[user_key] = value
-        end
-
-        def do_as(user, &block)
-          old = self.current
-
-          begin
-            self.current = user
-            response = block.call unless block.nil?
-          ensure
-            self.current = old
-          end
-
-          response
-        end
+        response
       end
     end
   end
+end
 end